feat(movie-card): enhance modal interaction with click-to-close feature
Introduce click-to-close functionality for the modal in MovieCardComponent. A new onModalClick method was added to close the modal when clicking outside its content. Additionally, ElementRef is used instead of Renderer2 for selecting modals, improving efficiency. In ModalDirective, the visibility toggle now automatically focuses on the modal when it's shown, enhancing accessibility. Landing component updates with a loading placeholder for the FAQ section while content is being fetched. These changes improve user experience and maintainability.